refactor: reorganize LLM configuration and provider structure, removing unused files and consolidating imports
