--- app/AI/llm_config/groq_config.py
@@ -1,8 +1,0 @@
-from app.config import GROQ_API_KEY
-from app.AI.llm_config._base_config import LLMModelConfig
-
-
-class GROQConfig(LLMModelConfig):
-    @property
-    def api_key(self) -> str:
-        return GROQ_API_KEY
